+++ src/test/java/io/nagurea/smsupsdk/contacts/remove/RemoveContactServiceIntTest.java
@@ -1,14 +1,12 @@
-package io.nagurea.smsupsdk.contacts.deduplicate;
+package io.nagurea.smsupsdk.contacts.remove;
 
 import io.nagurea.smsupsdk.common.status.ResponseStatus;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockserver.configuration.ConfigurationProperties;
 import org.mockserver.integration.ClientAndServer;
 import org.mockserver.model.HttpResponse;
-import org.mockserver.model.HttpStatusCode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
@@ -20,39 +18,39 @@ import static org.mockserver.model.HttpRequest.request;
 
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = SpringConfiguration.class)
-class ClearListServiceTest {
+class RemoveContactServiceIntTest {
+
     private static final String YOUR_TOKEN = "Your Token";
     private static final String EXPECTED_TOKEN = "Bearer " + YOUR_TOKEN;
-    private static final String LIST_ID = "2";
+    private static final String CONTACT_ID = "5a0331bffc5886074551ce97";
 
     /**
-     * Useless. Only here to show how services could be used with Spring
+     * Useless. Only here to see how services could be used with Spring
      */
     @Autowired
-    private DeduplicateService deduplicateService;
+    private RemoveContactService removeContactService;
 
     private static ClientAndServer mockServer;
 
     @BeforeAll
-    public static void startMockSMSUpServer() {
-        ConfigurationProperties.logLevel("DEBUG");
+    public static void startMockSMSUpServer(){
         mockServer = ClientAndServer.startClientAndServer("localhost", 4242, 4242);
         mockServer.when(
                 request()
-                        .withPath("/list/deduplicate/" + LIST_ID )
-                        .withMethod("PUT")
+                        .withPath("/list/contact/" + CONTACT_ID)
+                        .withMethod("DELETE")
                         .withHeader("Authorization", EXPECTED_TOKEN)
         ).respond(
                 HttpResponse.response()
-                        .withStatusCode(HttpStatusCode.OK_200.code())
+                        .withStatusCode(200)
                         .withBody(
                                 "{\n" +
                                         "  \"status\": 1,\n" +
-                                        "  \"message\": \"OK\",\n" +
-                                        "  \"removed\": 4\n" +
+                                        "  \"message\": \"OK\"\n" +
                                         "}"
                         )
         );
+
     }
 
     @AfterAll
@@ -60,26 +58,26 @@ class ClearListServiceTest {
         mockServer.stop();
     }
 
-    @Test
-    void deduplicate() throws IOException {
-        //given
-        //expected results
-        final DeduplicateResultResponse expectedResponse = DeduplicateResultResponse.builder()
-                .message(ResponseStatus.OK.getDescription())
-                .removed(4)
-                .build();
-        final int expectedStatusCode = 200;
+     @Test
+     void removeContact() throws IOException {
+         //given
+         final RemoveContactResultResponse expectedResponse = RemoveContactResultResponse.builder()
+                 .message(ResponseStatus.OK.getDescription())
+                 .responseStatus(ResponseStatus.OK)
+                 .build();
+         final int expectedStatusCode = 200;
 
-        //given arguments
+         //when
+         final RemoveContactResponse result = removeContactService.removeContact(YOUR_TOKEN, CONTACT_ID);
+         final Integer effectiveStatusCode = result.getStatusCode();
+         final RemoveContactResultResponse effectiveResponse = result.getEffectiveResponse();
 
-        //when
-        final DeduplicateResponse result = deduplicateService.deduplicate(YOUR_TOKEN, LIST_ID);
-        final Integer effectiveStatusCode = result.getStatusCode();
-        final DeduplicateResultResponse effectiveResponse = result.getEffectiveResponse();
+         //then
+         assertEquals(expectedStatusCode, effectiveStatusCode);
+         assertEquals(expectedResponse, effectiveResponse);
+         assertEquals(expectedResponse.getMessage(), effectiveResponse.getMessage());
+         assertEquals(expectedResponse.getStatus(), effectiveResponse.getStatus());
 
-        //then
-        assertEquals(expectedStatusCode, effectiveStatusCode);
-        assertEquals(expectedResponse, effectiveResponse);
     }